Feature/store impl final

+++ src/com/lambton/store/user/UserStore.java
@@ -8,5 +8,5 @@ import java.util.List;
 import java.util.Optional;
 
 public interface UserStore<T extends User> extends Store<T> {
-    List<User> search(int page, int size, Optional<String> firstName, Optional<String> lastName);
+    List<User> search(long page, long size, Optional<String> firstName, Optional<String> lastName);
 }
